--- backend/demo/src/main/java/com/example/demo/Bus.java
@@ -1,16 +1,0 @@
-package com.example.demo;
-
-import jakarta.persistence.*;
-
-//This annotation basically allows for each row entry in the database
-//We should figure out what should be the type of the rows
-//Buses or something else (thinking buses because the api result is only going to pop up data for that stop so the row type
-// should be what we are measuring or keeping track of at each stop which is the buses that are reaching it ) 
-@Entity
-public class Bus {
-
-    //Object variables line up with sql columns
-
-    //constructor so th
-    
-}
