--- backend/demo/src/main/java/com/example/demo/BusDB.java
@@ -1,5 +1,0 @@
-package com.example.demo;
-
-public class BusDB {
-    
-}
